+++ packages/Card/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/telus/pm-kit/compare/@pm-kit/card@1.0.0...@pm-kit/card@1.1.0) (2020-01-27)
+
+
+### Features
+
+* **card:** updated card styling and documentation ([#39](https://github.com/telus/pm-kit/issues/39)) ([53a4dba](https://github.com/telus/pm-kit/commit/53a4dba292c107d3da9218a83dd8086ae01ecb5f))
+
+
+
+
+
 # 1.0.0 (2020-01-24)
 
 **Note:** Version bump only for package @pm-kit/card