+++ packages/Dropdown/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/telus/pm-kit/compare/@pm-kit/dropdown@1.0.2...@pm-kit/dropdown@1.0.3) (2020-02-03)
+
+
+### Bug Fixes
+
+* **dropdown:** added a wrapper div around label and dropdown ([#51](https://github.com/telus/pm-kit/issues/51)) ([843f9c3](https://github.com/telus/pm-kit/commit/843f9c35d5297a8391e7fbd0fd8891028323b98a))
+
+
+
+
+
+## [1.0.2](https://github.com/telus/pm-kit/compare/@pm-kit/dropdown@1.0.1...@pm-kit/dropdown@1.0.2) (2020-01-31)
+
+
+### Bug Fixes
+
+* **dropdown:** fixed window undefined bug from react-select ([#50](https://github.com/telus/pm-kit/issues/50)) ([0eaa9b3](https://github.com/telus/pm-kit/commit/0eaa9b35abb76997db39e160814fe9ce1fe1bbb0))
+
+
+
+
+
 ## [1.0.1](https://github.com/telus/pm-kit/compare/@pm-kit/dropdown@1.0.0...@pm-kit/dropdown@1.0.1) (2020-01-24)
 
 **Note:** Version bump only for package @pm-kit/dropdown